% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{CalcError}
\alias{CalcError}
\title{Reconstruction error}
\usage{
CalcError(measurement, reconstruction, error = "SoS")
}
\arguments{
\item{measurement}{data matrix of the original measurement}

\item{reconstruction}{data matrix of the reconstructed signals}

\item{error}{the type of error to calculate. currently supported: SoS, RMSD, MAE, MAPE, CosineSimilarity, WassersteinDistance}
}
\value{
vector with the reconstruction error for each row
}
\description{
Calculates the reconstruction error of the unmixing. Current options for the error calculation are:
}
\details{
- Sum of squares (SoS)
- Root mean square deviation (RMSD)
- Mean absolute error (MAE)
- Mean absolute percentage error (MAPE)
- Cosine distance (CosineDistance) 1 - Cosine similarity
- Wasserstein distance (WassersteinDistance) Note: slow compared to the others
}
