% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{CalcMedianSpectra}
\alias{CalcMedianSpectra}
\title{Calculate reference spectra by median subtraction}
\usage{
CalcMedianSpectra(pos, neg = NULL, channels = NULL, force_non_negative = T)
}
\arguments{
\item{pos}{Positive population(s) medians supplied as dataframe containing the channel medians in the columns. Can contain multiple rows.}

\item{neg}{Negative population medians supplied as dataframe containing the channel medians in the columns. Can contain only one row. If no negative population is supplied, only channel normalization is done (e.g., for autofluorescence populations).}

\item{channels}{Optional, the channels (column names) to include in the calculation.}

\item{force_non_negative}{TRUE (default) / FALSE. If TRUE, sets all negative values to 0. Careful: negative values (if not very small) usually indicate a mismatch between the positive and negative population autofluorescence!}
}
\value{
A tibble containing the reference spectra (channel values in columns) together with file column taken from pos.
}
\description{
Classical way of calculating reference spectra by subtracting the medians of a negative population from the medians of a positive population and then normalizing to the highest channel reading.
Multiple dyes can be processed in one go, as long as they share the same negative population.
}
