% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{ChannelCheck}
\alias{ChannelCheck}
\title{Check if channel list is given and present in data}
\usage{
ChannelCheck(dat, channels)
}
\arguments{
\item{dat}{tibble, output of GetData or GetMedianData}

\item{channels}{character vector, channels to be used.}
}
\value{
list of channels to be used
}
\description{
Internal function
}
