% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_afClustering.R
\name{ClusterSelection}
\alias{ClusterSelection}
\title{Select Autofluorescence Clusters}
\usage{
ClusterSelection(
  dat,
  clusters,
  channels = NULL,
  count_min = NULL,
  csim_max = NULL,
  selection = "max"
)
}
\arguments{
\item{dat}{data read in using GetData}

\item{clusters}{vector with cluster assignment for each entry in dat}

\item{channels}{character vector, channel selection to include in the plot (default: NULL, i.e. all)}

\item{count_min}{integer, minimum number of events needed for a cluster to be considered in the selection (default: NULL, i.e. all)}

\item{csim_max}{numeric, maximum cosine similarity value. Clusters with a higher cosine similarity value will be combined. (default: NULL, i.e. no combining of any clusters)}

\item{selection}{either "max" or "median". If set to "max", the cluster with the highest intensity is chosen from a group of similar clusters as spectrum. If set to "median", the median of all events within a group of similar clusters is taken as spectrum.}
}
\value{
List of outputs. The spectra element contains the dataframe with the selected spectra.
}
\description{
For a clustering of autofluorescence populations, find similar clusters based on the cosine similarity.
}
