% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{CosSim}
\alias{CosSim}
\title{Calculate cosine similarity matrix}
\usage{
CosSim(m)
}
\arguments{
\item{m}{Matrix}
}
\value{
similarity matrix
}
\description{
https://stats.stackexchange.com/questions/31565/compute-a-cosine-dissimilarity-matrix-in-r
}
