% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{GetData}
\alias{GetData}
\title{Read data from FCS file into tibble or convert flowFrame into tibble}
\usage{
GetData(f, pattern = NA, file_name = NA, ...)
}
\arguments{
\item{f}{Path to FCS file(s), flowFrame or flowSet}

\item{pattern}{Regular expression to extract pattern from file name (default: NA)}

\item{file_name}{File name(s) to be stored in tibble}

\item{...}{Arguments passed to read.FCS}
}
\value{
Tibble with the exprs matrix of the FCS file and a column containing the file name or extracted pattern. If both pattern and file_name are NA, the literal filename will be used.
}
\description{
Read data from FCS file into tibble or convert flowFrame into tibble
}
