% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{GetMedianData}
\alias{GetMedianData}
\title{Get channel medians for FCS file or flowFrame}
\usage{
GetMedianData(f, pattern = NULL, file_name = NULL, ...)
}
\arguments{
\item{f}{Path to FCS file(s), flowFrame or flowSet, or output of GetData}

\item{pattern}{Regular expression to extract pattern from file name (default: NA)}

\item{file_name}{File name to be stored in tibble}

\item{...}{Arguments passed to read.FCS}
}
\value{
Tibble with the channel medians in the columns with additional file column.
}
\description{
Get channel medians for FCS file or flowFrame
}
