% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{MakeM}
\alias{MakeM}
\title{Create unmixing matrix}
\usage{
MakeM(dye_spectra, channels = NA)
}
\arguments{
\item{dye_spectra}{Dataframe containing the dye spectra in the rows.}

\item{channels}{Optional, vector containing the channel names to be used. Default (NA) uses all channels.}
}
\value{
Unmixing matrix
}
\description{
Create unmixing matrix
}
