% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{MakeMList}
\alias{MakeMList}
\title{Create list of unmixing matrices}
\usage{
MakeMList(AF, dye_spectra, channels = NULL)
}
\arguments{
\item{AF}{Dataframe containing the autofluorescence spectra in the rows.}

\item{dye_spectra}{Dataframe containing the dye spectra in the rows.}

\item{channels}{Optional, vector containing the channel names to be used. Default (NA) uses all channels.}
}
\value{
List of unmixing matrices to used in Unmixing
}
\description{
Creates a list of unmixing matrices to be used when there are multiple autofluorescence populations and each event should be unmixed using only one autofluorescence population. Each unmixing matrix contains the dye spectra plus one autofluorescence spectra.
}
