% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_viz.R
\name{PlotMxN}
\alias{PlotMxN}
\title{Plot one marker against all others}
\usage{
PlotMxN(marker, dat_unmixed, all_markers, output = "screen", output_path = ".")
}
\arguments{
\item{marker}{string, the marker to be plotted agains all other markers}

\item{dat_unmixed}{dataframe, the unmixed data to be plotted. Can contain data from multiple single stains.}

\item{all_markers}{vector, all markers to be included in the plot}

\item{output}{either "screen", "file" or "both". Whether to return the plot ("screen"), write it to a png ("file") or do both}

\item{output_path}{string, where to save the plot if output is set to either "file" or "both". Default: ".", the script directory.}
}
\value{
either a ggplot2 plot, the plot as a png file or both
}
\description{
Controls the unmixing quality by plotting one marker against all others.
}
\details{
Best done on data from single stained cells that was unmixed with a given unmixing matrix. If the plots show any curved distributions, there is a problem with the unmixing (e.g., a single stain needs to be recorded on cells instead of beads).
}
