% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{PlotRefSpectra}
\alias{PlotRefSpectra}
\title{Plot reference spectra}
\usage{
PlotRefSpectra(spectra, transformation = NA)
}
\arguments{
\item{spectra}{spectra dataframe as produced by MedianSpectra or RegSpectra}

\item{transformation}{optional asinh transform of the spectra. Default: NA. If you want to transform the normalized spectra, pass the cofactor of the asinh transformation. Note: The spectra are max normalized to 1 so your usual cofactors will not work here. Try cofactors around 0.001.}
}
\value{
ggplot object
}
\description{
Plot reference spectra
}
