% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_afClustering.R
\name{PlotScatterClusters}
\alias{PlotScatterClusters}
\title{Plot scatter signals of AF clusters}
\usage{
PlotScatterClusters(
  input,
  scatter_channels = c("FSC-A", "SSC-A"),
  transformation = NULL,
  bins = 100,
  alpha = 0.5,
  ncol = 3
)
}
\arguments{
\item{input}{result of ClusterSelection}

\item{scatter_channels}{character vector, names of the two scatter channels to plot (default: c("FSC-A", "SSC-A"))}

\item{transformation}{numeric, cofactor used for the asinh transformation (default: NULL, i.e., none)}

\item{bins}{numeric, number of bins to use in the 2D histograms (default: 100)}

\item{alpha}{numeric, transparency parameter of the 2D histograms (default: 0.5)}

\item{ncol}{integer, number of columns of the facet plot (default: NULL, i.e. automatic)}
}
\value{
ggplot object, showing the scatter signals for each group of clusters (colored by cluster) found by ClusterSelection. If a group labelled NA is shown in the plot, clusters were excluded from the analysis in ClusterSelection for having too few events.
}
\description{
Plot scatter signals of AF clusters
}
