% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{PlotSpectra}
\alias{PlotSpectra}
\title{Plot Spectra}
\usage{
PlotSpectra(
  dat,
  channels = NULL,
  bins = 50,
  transformation = 1000,
  ncol = NULL
)
}
\arguments{
\item{dat}{tibble, output of GetData}

\item{channels}{character vector, channel selection to include in the plot (default: NULL, i.e. all)}

\item{bins}{numeric, number of bins to use in the channel heatmaps (default: 50)}

\item{transformation}{numeric, cofactor used for the asinh transformation (default: 1000)}

\item{ncol}{integer, number of columns of the facet plot (default: NULL, i.e. automatic)}
}
\value{
ggplot object
}
\description{
Plot Spectra
}
