% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_afClustering.R
\name{PlotSpectraClusters}
\alias{PlotSpectraClusters}
\title{Plot clustered AF spectra}
\usage{
PlotSpectraClusters(input, transformation = 0.1, bins = 50, ncol = 3)
}
\arguments{
\item{input}{result of ClusterSelection}

\item{transformation}{numeric, cofactor used for the asinh transformation (default: 0.1) Note: The spectra are max normalized to 1 so your usual cofactors will not work here.}

\item{bins}{numeric, number of bins to use in the channel heatmaps (default: 50)}

\item{ncol}{integer, number of columns of the facet plot (default: NULL, i.e. automatic)}
}
\value{
ggplot object
}
\description{
Plot clustered AF spectra
}
