% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{RegSpectra}
\alias{RegSpectra}
\title{Calculate reference spectra by regression}
\usage{
RegSpectra(
  pos,
  channels = NULL,
  neg = NULL,
  force_non_negative = T,
  robust = T
)
}
\arguments{
\item{pos}{Positive population(s) medians supplied as dataframe containing the channel medians in the columns. Can contain multiple rows.}

\item{channels}{Optional, the channels (column names) to include in the calculation.}

\item{neg}{Optional, negative population medians supplied as dataframe containing the channel medians in the columns. Can contain only one row. Needed if pos contains to few negative events.}

\item{force_non_negative}{TRUE (default) / FALSE. If TRUE, sets all negative values to 0. Careful: negative values (if not very small) usually indicate a mismatch between the positive and negative population autofluorescence!}

\item{robust}{TRUE (default) / FALSE. If TRUE, rebust regression is done via MASS::rlm, otherwise standard regression is done using lm.}
}
\value{
A tibble containing the reference spectra (channel values in columns) together with file column taken from pos.
}
\description{
Calculates reference spectra based on regression between channels (like AutoSpill but without the additional optimization). Does not require pre-gating on positive and negative populations, but positive and negative events should ideally be present to similar amounts.
}
