% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{Unmix.ols}
\alias{Unmix.ols}
\title{Ordinary least squares unmixing}
\usage{
Unmix.ols(dat, M, bg = NULL)
}
\arguments{
\item{dat}{dataframe, data to be unmixed}

\item{M}{unmixing matrix}

\item{bg}{vector of constant background to be subtracted from each event}
}
\value{
dataframe with component abundances
}
\description{
https://en.wikipedia.org/wiki/Ordinary_least_squares
}
