% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{Unmix}
\alias{Unmix}
\title{Spectral unmixing}
\usage{
Unmix(
  data,
  M,
  bg = NULL,
  unmixing = "OLS",
  error = "SoS",
  write_FCS = FALSE,
  FCS_suffix = "unmixed",
  FCS_dir = ".",
  fullOutput = FALSE,
  n_cores = 1
)
}
\arguments{
\item{data}{Raw data to be unmixed. Can be passed as dataframe, tibble, flowframe or flowSet.}

\item{M}{Matrix containing the component spectra or list of such matrices.}

\item{bg}{Constant background or autofluorescence to be subtracted of events before unmixing (dataframe). Default: NULL}

\item{unmixing}{Unmixing method to use passed as string. Currently supported methods are: ordinary least squares ("OLS"),
weighted least squares ("WLS"), non-negative least squares ("NNLS"), weighted non-negative least squares ("WNNLS") and mean
absolute percentage error ("MAPE").}

\item{error}{Type of error to calculate for the unmixing result (NOT used during unmixing itself!).
The error will be added as an "error" column to the unmixed output and determines which autofluorescence population is chosen for the unmixing of each cell in case multiple AF populations are present (list of Ms given to function).
Choices: Sum of squares ("SoS"), Root mean square deviation ("RMSD"),
Mean absolute error ("MAE"), Mean absolute percentage error ("MAPE"),
Cosine distance ("CosineDistance"), Wasserstein distance ("WassersteinDistance", slow!)}

\item{write_FCS}{Write unmixed file to FCS file (TRUE / FALSE (default))?}

\item{FCS_suffix}{Suffix to be added to the unmixed FCS file name. Default: "unmixed".}

\item{FCS_dir}{Directory to write the unmixed FCS file to. Default: ".".}

\item{fullOutput}{If TRUE, returns a list containing the reconstructed data and unmixing matrix in addition to the abundances.
Necessary for the interactive visualization.}

\item{n_cores}{Number of cores to use during unmixing (testing the different unmixing matrices, default: 1). Only used if a list of unmixing marices is given as M.}
}
\value{
List containing the abundances (unmixed). If fullOutput = TRUE, the list also contains the reconstructed data (reconstruction),
and the unmixing matrix. If write_FCS = TRUE, the function also writes an FCS file containing the abundances.
}
\description{
Spectral unmixing using OLS, WLS, NNLS, WNNLS or MAPE with one or multiple autofluorescence populations.
For all weighted algorithms: For the weight of negative values, the absolute is taken and for 0 values, an offset of 0.05 is added.
}
