% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{Unmix.wrapper}
\alias{Unmix.wrapper}
\title{Wrapper for all defined unmixing methods}
\usage{
Unmix.wrapper(dat, M, bg = NULL, unmixFunc = "OLS")
}
\arguments{
\item{dat}{dataframe, data to be unmixed}

\item{M}{unmixing matrix}

\item{bg}{vector of constant background to be subtracted from each event (default: NULL)}

\item{unmixFunc}{string, unmixing method. OLS, WLS, NNLS, WNNLS or MAPE. (default: "OLS")}
}
\value{
matrix with component abundances
}
\description{
Wrapper for all defined unmixing methods
}
