% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFile}
\alias{UnmixFile}
\title{Unmix a dataset}
\usage{
UnmixFile(
  data,
  M,
  bg = NULL,
  unmixing = "OLS",
  write_FCS = FALSE,
  FCS_suffix = "unmixed",
  FCS_dir = ".",
  fullOutput = FALSE
)
}
\arguments{
\item{data}{dataframe to unmix}

\item{M}{unmixing matrix}

\item{bg}{dataframe containing background (optional)}

\item{unmixing}{unmixing function to use (string)}

\item{write_FCS}{write output into FCS files (default: FALSE)}

\item{FCS_suffix}{suffix added to the FCS files (default: "unmixed")}

\item{FCS_dir}{directory to export the FCS files to (default: ".")}

\item{fullOutput}{Add the reconstruction and unmixing matrix to the output (default: FALSE). Needed for the interactive visualization}
}
\value{
list(unmixed, reconstruction, M)
}
\description{
Use inside Unmix wrapper.
}
