% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFile_MultiAF}
\alias{UnmixFile_MultiAF}
\title{Unmixing with multiple AF}
\usage{
UnmixFile_MultiAF(
  data,
  M,
  bg = NULL,
  unmixing = "OLS",
  error = "SoS",
  write_FCS = FALSE,
  FCS_suffix = "unmixed",
  FCS_dir = ".",
  fullOutput = FALSE,
  n_cores = 1
)
}
\arguments{
\item{data}{dataframe to unmix}

\item{M}{list of unmixing matrices}

\item{bg}{dataframe containing background (optional)}

\item{unmixing}{unmixing function to use (string)}

\item{error}{type of error to calculate for the unmixing result (NOT used during unmixing itself!). Determines which AF population is used to unmix each cell. The error will be added as an "error" column to the unmixed output. Choices: Sum of squares ("SoS"), Root mean square deviation ("RMSD"), Mean absolute error ("MAE"), Mean absolute percentage error ("MAPE"), Cosine distance ("CosineDistance"), Wasserstein distance ("WassersteinDistance", slow!)}

\item{write_FCS}{write output into FCS files (default: FALSE)}

\item{FCS_suffix}{suffix added to the FCS files (default: "unmixed")}

\item{FCS_dir}{directory to export the FCS files to (default: ".")}

\item{fullOutput}{Add the reconstruction and unmixing matrix to the output (default: FALSE). Needed for the interactive visualization}

\item{n_cores}{Number of cores to use during unmixing (testing the different unmixing matrices, default: 1)}
}
\value{
list(unmixed, reconstruction, M)
}
\description{
Use multiple unmixing matrices, each with different AF and take best result for each cell.
AF columns in M list must have unique names (e.g. AF1, AF2, ...)!
}
\details{
Intended to be used inside Unmix wrapper.
}
