% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFunc.mape.bg}
\alias{UnmixFunc.mape.bg}
\title{MAPE unmixing with backround}
\usage{
UnmixFunc.mape.bg(x, M, bg)
}
\arguments{
\item{x}{event row}

\item{M}{unmixing matrix}

\item{bg}{vector of constant background to be subtracted from each event}
}
\value{
vector with component abundances
}
\description{
Use in wrapper
question of how to treat negative values and whether to subtract background for the weights (probably not).
}
