% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFunc.mape.nobg}
\alias{UnmixFunc.mape.nobg}
\title{MAPE unmixing without backround}
\usage{
UnmixFunc.mape.nobg(x, M)
}
\arguments{
\item{x}{event row}

\item{M}{unmixing matrix}
}
\value{
vector with component abundances
}
\description{
Use in wrapper
question of how to treat negative values and whether to subtract background for the weights (probably not).
}
