% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFunc.nnls}
\alias{UnmixFunc.nnls}
\title{Non-negative least squares unmixing}
\usage{
UnmixFunc.nnls(x, M)
}
\arguments{
\item{x}{event row}

\item{M}{unmixing matrix}
}
\value{
vector with component abundances
}
\description{
Non-negative least squares unmixing
}
