% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFunc.ols}
\alias{UnmixFunc.ols}
\title{Ordinary least squares unmixing}
\usage{
UnmixFunc.ols(x, U)
}
\arguments{
\item{x}{event row}

\item{U}{(M^TM)*M^T}
}
\value{
abundance matrix
}
\description{
Ordinary least squares unmixing
}
