% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFunc.wls.bg}
\alias{UnmixFunc.wls.bg}
\title{Weighted least squares}
\usage{
UnmixFunc.wls.bg(x, M, bg)
}
\arguments{
\item{x}{event row}

\item{M}{unmixing matrix}

\item{bg}{vector of constant background to be subtracted from each event}
}
\value{
vector with component abundances
}
\description{
Weighted least squares
}
