% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFunc.wls.nobg}
\alias{UnmixFunc.wls.nobg}
\title{Weighted least squares}
\usage{
UnmixFunc.wls.nobg(x, M)
}
\arguments{
\item{x}{event row}

\item{M}{unmixing matrix}
}
\value{
vector with component abundances
}
\description{
Weighted least squares
}
