% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFunc.wnnls.bg}
\alias{UnmixFunc.wnnls.bg}
\title{Weighted NNLS with background}
\usage{
UnmixFunc.wnnls.bg(x, M, bg)
}
\arguments{
\item{x}{event row}

\item{M}{unmixing matrix}

\item{bg}{vector of constant background to be subtracted from each event}
}
\value{
vector with component abundances
}
\description{
Weighted NNLS with background
}
