% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixFunc.wnnls.nobg}
\alias{UnmixFunc.wnnls.nobg}
\title{Weighted NNLS without background}
\usage{
UnmixFunc.wnnls.nobg(x, M)
}
\arguments{
\item{x}{event row}

\item{M}{unmixing matrix}
}
\value{
vector with component abundances
}
\description{
Weighted NNLS without background
}
