% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_viz.R
\name{UnmixingCompViz}
\alias{UnmixingCompViz}
\title{Compare two unmixing results side by side}
\usage{
UnmixingCompViz(unmixing1, unmixing2, original, n_sample = 1000)
}
\arguments{
\item{unmixing1}{first unmixing result (output of Unmix with full_output = TRUE)}

\item{unmixing2}{second unmixing result (output of Unmix with full_output = TRUE)}

\item{original}{original raw data (output of GetData)}

\item{n_sample}{number of data points to sample for plots}
}
\value{
opens interactive visualization
}
\description{
The top plot shows the abundances resulting from the unmixing in a 2D dot plot. The user can choose which parameters to display and how to transform them. The color indicates the unmixing error.
By clicking on a data point, details for that point are shown in the plots below. In the middle plots, the fitted spectral contributions for the selected data point are shown. The bottom plot shows the position of the selected data point in FSC/SSC.
}
