% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixMatrix}
\alias{UnmixMatrix}
\title{Wrapper to apply unmixing function to each row in data matrix}
\usage{
UnmixMatrix(dat, UnmixFun, ...)
}
\arguments{
\item{dat}{Dataframe to unmix}

\item{UnmixFun}{unmixing function to apply}

\item{...}{Parameters passed to unmixing function}
}
\value{
abundance matrix
}
\description{
Wrapper to apply unmixing function to each row in data matrix
}
