% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{UnmixMatrix_par}
\alias{UnmixMatrix_par}
\title{Wrapper to apply unmixing function to each row in data matrix (parallel version)}
\usage{
UnmixMatrix_par(n_cores, dat, UnmixFun, ...)
}
\arguments{
\item{n_cores}{Number of cores to use}

\item{dat}{Dataframe to unmix}

\item{UnmixFun}{unmixing function to apply}

\item{...}{Parameters passed to unmixing function}
}
\value{
abundance matrix
}
\description{
Wrapper to apply unmixing function to each row in data matrix (parallel version)
}
