% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{SampleCorr}
\alias{SampleCorr}
\title{Correct for background}
\usage{
SampleCorr(dat, neg_median = NULL, channels = NULL)
}
\arguments{
\item{dat}{dataframe containing raw events}

\item{neg_median}{dataframe containing the autofluorescence to subtract}

\item{channels}{optional, channels to use}
}
\value{
tibble with the corrected sample
}
\description{
Used in SingleStainVariability
}
