% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_SingleStains.R
\name{SingleStainVariability}
\alias{SingleStainVariability}
\title{Plot the variability of the single stains}
\usage{
SingleStainVariability(
  input_list,
  spectra,
  channels = NULL,
  transformation = 0.01,
  bins = 50,
  ncol = NULL
)
}
\arguments{
\item{input_list}{list of lists in the form of list(list(pos_raw, neg_median), ...). Each sublist must contain pos_raw, the dataframe of the single stain raw data and neg_median, the channel medians of the corresponding negative control.}

\item{spectra}{tibble, the calculated reference spectra (output of MedianSpectra or RegSpectra).}

\item{channels}{character vector, channels to be included in the calculation in calculation (default: NULL, i.e., all)}

\item{transformation}{numeric, optional asinh transform of the spectra. Default: 0.01 If you want to transform the normalized spectra, pass the cofactor of the asinh transformation. Note: The spectra are max normalized to 1 so your usual cofactors will not work here. Try cofactors around 0.01.}

\item{bins}{numeric, number of bins to use in the channel heatmaps (default: 50)}

\item{ncol}{integer, number of columns of the facet plot (default: NULL, i.e. automatic)}
}
\value{
Plot of single stain variability, faceted by single stain. The red line shows the calculated reference spectra, the blue lines the (autofluorescence subtracted) single stain events. The blue shade indicates the average similarity of that event to all other events (light blue: more similar, dark blue: less dissimilar).
}
\description{
Plot the variability of the single stains
}
