% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{Unmix.ols}
\alias{Unmix.ols}
\title{Ordinary least squares unmixing}
\usage{
Unmix.ols(dat, M, bg = NULL, n_cores = 1)
}
\arguments{
\item{dat}{Dataframe, data to be unmixed}

\item{M}{Unmixing matrix}

\item{bg}{Vector of constant background to be subtracted from each event}

\item{n_cores}{Number of cores to use during unmixing (default: 1)}
}
\value{
dataframe with component abundances
}
\description{
https://en.wikipedia.org/wiki/Ordinary_least_squares
}
