% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unmixing_Methods.R
\name{Unmix.wnnls}
\alias{Unmix.wnnls}
\title{Weighted NNLS unmixing}
\usage{
Unmix.wnnls(dat, M, bg = NULL, n_cores = 1)
}
\arguments{
\item{dat}{dataframe, data to be unmixed}

\item{M}{unmixing matrix}

\item{bg}{vector of constant background to be subtracted from each event}

\item{n_cores}{Number of cores to use during unmixing (default: 1)}
}
\value{
dataframe with component abundances
}
\description{
https://stackoverflow.com/questions/47888996/weighted-nonnegative-least-squares-in-r
}
